+++ src/webview_provider/lineagePanel.ts
@@ -20,7 +20,6 @@ import {
 } from "../manifest/event/manifestCacheChangedEvent";
 import { ModelGraphViewPanel } from "./modelGraphViewPanel";
 import { NewLineagePanel } from "./newLineagePanel";
-import { inject } from "inversify";
 
 export interface LineagePanelView extends WebviewViewProvider {
   init(): void;
